Extracted Shrink from GenTree
Taken verbatim from:

https://github.com/hedgehogqa/haskell-hedgehog/blob/47b79adb33ed1d7098624a6528e1e13aa741ba59/src/Hedgehog/Internal/Shrink.hs
